Break refractor functionality into generalized functionality
The refractor functionality included getting, checking, modifying tags, and updating objects. This will be almost the same process for removing and adding tags. Broke the existing functionality into more generalized functions to be able to reuse when implementing the remove tag functionality. This includes a rewrite of the handle_refactor_tags() function to call the functionality that was moved to separate functions.